// src/app/models/cotizacion/cotizacion.model.ts
export class cotizaciones {
  id: number;  
  Nombres: string;
  Apellidos: string;
  Telefono: string;
  Correo: string;
  Direccion: string;
  Observaciones: string;
  ContactoVitalea: number;
  CotizacionWhatsapp: number;
  IdUsuario: number;
  plan: number;
  Examenes: any;
  Chequeos: any;
  Asesor: string;
  FechaCotizacion: string;
  Tipo:number;
  codigo:string;
  procedimiento:string;
  valor:string;
}