// src/app/models/Generic/Generic.model.ts
export class GenericCity {
  id: number;
  ciudad: string;
}

export class GenericDepartament {
  id: number;
  depto: string;
}

export class GenericCountry {
  id: number;
  pais: string;
}

export class GenericLacation {
  id: number;
  localidad: string;
}

export class GenericBarrio {
  id: number;
  barrio: string;
}

export class GenericTypeAfil {
  id: number;
  tipoafiliacion: string;
}

export class GenericEstrato {
  id: number;
  estrato: string;
}

export class GenericGenero {
  id: number;
  genero: string;
}

export class GenericTipoDoc {
  id: number;
  tipodocumento: string;
}

export class GenericEstadoCivil {
  id: number;
  estadoCiv: string;
}

export class GenericEps {
  id: number;
  eps: string;
}

export class GenericCondition {
  id: number;
  estado: string;
}

export class Options {
  IdOpc: number;
  Opcion: string;
  Icono: string;   
}